refactor: utils cleanup
