+++ packages/effect-ext-utils/README.md
@@ -10,7 +10,7 @@ This package provides reusable utilities and helper functions for building VS Co
 
 1. This is a package, you can import it directly instead of having to get it through the extension API
 1. it's a dev-time, not run-time dependency
-1. each extension get its own instance of this package, so they aren't shared/comingled. Ex: `extensionScope` is per extension, each extension manages and closes its own scope.
+1. each extension get its own instance of this package, so they aren't shared/commingled. Ex: `extensionScope` is per extension, each extension manages and closes its own scope.
 1. You can pass these to `services` as dependencies (ex: some Effect that requires a scope).
 1. the dependencies are minimal (mostly Effect, which all extensions will end up with). This should **not** contain any DX libraries or large dependencies
