+++ packages/salesforcedx-lightning-lsp-common/tsconfig.json
@@ -2,6 +2,7 @@
   "extends": "../../tsconfig.common.json",
   "compilerOptions": {
     "outDir": "./out",
+    "rootDir": ".",
     "strictNullChecks": true,
     "skipLibCheck": true
   },
