+++ packages/salesforcedx-utils-vscode/src/helpers/utils.ts
@@ -93,19 +93,6 @@ export const stripAnsiInJson = (str: string, hasJson: boolean): string => (str &
 
 export const stripAnsi = (str: string): string => (str ? str.replaceAll(ansiRegex(), '') : str);
 
-export const getMessageFromError = (err: any): string => {
-  if (err instanceof Error) {
-    return err.message;
-  }
-  if (typeof err === 'string') {
-    return err;
-  }
-  if (err) {
-    return `Unexpected error: ${JSON.stringify(err)}`;
-  }
-  return 'Unknown error';
-};
-
 /*
 Copied from https://github.com/sindresorhus/strip-ansi/blob/master/index.js
