+++ packages/salesforcedx-utils-vscode/test/jest/helpers/fs.test.ts
@@ -13,10 +13,7 @@ import {
   createDirectory,
   readDirectory,
   stat,
-  safeDelete,
-  isDirectory,
-  isFile,
-  rename
+  safeDelete
 } from '../../../src/helpers/fs';
 
 jest.mock('vscode');
@@ -92,52 +89,6 @@ describe('file system utilities', () => {
     });
   });
 
-  describe('isDirectory', () => {
-    it('should return true when path is a directory', async () => {
-      (vscode.workspace.fs.stat as jest.Mock).mockResolvedValue({ type: vscode.FileType.Directory });
-
-      const result = await isDirectory('/test/path');
-      expect(result).toBe(true);
-    });
-
-    it('should return false when path is a file', async () => {
-      (vscode.workspace.fs.stat as jest.Mock).mockResolvedValue({ type: vscode.FileType.File });
-
-      const result = await isDirectory('/test/path');
-      expect(result).toBe(false);
-    });
-
-    it('should return false when path does not exist', async () => {
-      (vscode.workspace.fs.stat as jest.Mock).mockRejectedValue(mockError);
-
-      const result = await isDirectory('/test/path');
-      expect(result).toBe(false);
-    });
-  });
-
-  describe('isFile', () => {
-    it('should return true when path is a file', async () => {
-      (vscode.workspace.fs.stat as jest.Mock).mockResolvedValue({ type: vscode.FileType.File });
-
-      const result = await isFile('/test/path');
-      expect(result).toBe(true);
-    });
-
-    it('should return false when path is a directory', async () => {
-      (vscode.workspace.fs.stat as jest.Mock).mockResolvedValue({ type: vscode.FileType.Directory });
-
-      const result = await isFile('/test/path');
-      expect(result).toBe(false);
-    });
-
-    it('should return false when path does not exist', async () => {
-      (vscode.workspace.fs.stat as jest.Mock).mockRejectedValue(mockError);
-
-      const result = await isFile('/test/path');
-      expect(result).toBe(false);
-    });
-  });
-
   describe('createDirectory', () => {
     it('should create directory if it does not exist', async () => {
       (vscode.workspace.fs.createDirectory as jest.Mock).mockResolvedValue(undefined);
@@ -172,26 +123,6 @@ describe('file system utilities', () => {
       (vscode.workspace.fs.createDirectory as jest.Mock).mockRejectedValue(mockError);
 
       await expect(createDirectory('/test/path')).rejects.toThrow('Failed to create directory /test/path: Test error');
-    });
-  });
-
-  describe('rename', () => {
-    it('should rename file successfully', async () => {
-      (vscode.workspace.fs.rename as jest.Mock).mockResolvedValue(undefined);
-
-      await rename('/old/path/file.txt', '/new/path/file.txt');
-      expect(vscode.workspace.fs.rename).toHaveBeenCalledWith(
-        { fsPath: '/old/path/file.txt' },
-        { fsPath: '/new/path/file.txt' }
-      );
-    });
-
-    it('should throw error when rename fails', async () => {
-      (vscode.workspace.fs.rename as jest.Mock).mockRejectedValue(mockError);
-
-      await expect(rename('/old/path/file.txt', '/new/path/file.txt')).rejects.toThrow(
-        'Failed to rename /old/path/file.txt to /new/path/file.txt: Test error'
-      );
     });
   });
