+++ packages/salesforcedx-utils-vscode/test/jest/telemetry/reporters/telemetryFile.test.ts
@@ -5,7 +5,7 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import * as vscode from 'vscode';
-import * as workspaceUtils from '../../../../src';
+import * as workspaceUtils from '../../../../src/workspaces/workspaceUtils';
 import { TelemetryFile } from '../../../../src/telemetry/reporters/telemetryFile';
 
 jest.mock('vscode');
