+++ packages/salesforcedx-utils/package.json
@@ -82,9 +82,6 @@
       ]
     }
   },
-  "dependencies": {
-    "@salesforce/core": "^8.31.0"
-  },
   "devDependencies": {
     "rxjs": "^7.8.2"
   }
